Fixed comments from PR #60
