+++ backend/nginx/dev.nginx.conf
@@ -10,8 +10,6 @@ http {
   server {
     listen 80;
     listen [::]:80;
-    listen 8080;
-    listen [::]:8080;
 
     location /socket.io/ {
       proxy_pass http://backend/socket.io/;
