+++ backend/src/validation/schemas.ts
@@ -3,10 +3,10 @@ import {z, type ZodType} from "zod";
 
 export const placeholderSchema = z.string() satisfies ZodType<Placeholder>;
 
-const passWordSchema = z
+const passwordSchema = z
 	.string()
 	.min(8, "Password must be at least 8 characters long")
-	.refine((password) => /[a-z]/.test(password), "Password must contain at least one small letter")
+	.refine((password) => /[a-z]/.test(password), "Password must contain at least 1 small letter")
 	.refine((password) => /[A-Z]/.test(password), "Password must contain at least 1 capital letter")
 	.refine((password) => /[0-9]/.test(password), "password must contain at least 1 number")
 	.refine((password) => /[!@#$%^&*]/.test(password), "Password must contain at least 1 special character '!@#$%^&*'");
@@ -17,10 +17,10 @@ const usernameSchema = z
 	.max(20, "Username can't be over 20 characters long")
 	.regex(/^[a-zA-Z0-9_]+$/, "Username must contain only small/capital letter, numbers or underscore");
 
-export const UserSignupSchema = z.object({
+export const SignupSchema = z.object({
 	username: usernameSchema,
 	email: z.email(),
-	password: passWordSchema,
+	password: passwordSchema,
 }) satisfies ZodType<SigningUser>;
 
 // id here for testing
